:sparkles: merge clair and clair klar together
