+++ docs/content/en/integrations/parsers/file/clair.md
@@ -2,4 +2,7 @@
 title: "Clair Scan"
 toc_hide: true
 ---
-Import JSON reports of Docker image vulnerabilities.
+Import JSON reports of Docker image vulnerabilities from clair or the clair klar client.
+
+### Sample Scan Data
+Sample Clair and Clair Klar scans can be found [here](https://github.com/DefectDojo/django-DefectDojo/tree/master/unittests/scans/clair).
