--- docs/content/en/integrations/parsers/file/clair_klar.md
@@ -1,6 +1,0 @@
----
-title: "Clair Klar Scan"
-toc_hide: true
----
-Import JSON reports of Docker image vulnerabilities from clair klar
-client.
