+++ unittests/scans/clair/clair_empty.json
