+++ unittests/scans/clair/clair_few_vuln.json
