+++ unittests/scans/clair/clairklar_high.json
