+++ unittests/scans/clair/clairklar_mixed.json
