+++ unittests/tools/test_clair_parser.py
@@ -4,15 +4,22 @@ from dojo.tools.clair.parser import ClairParser
 
 class TestClairParser(DojoTestCase):
 
-    def test_no_findings(self):
-        my_file_handle = open("unittests/scans/clair/empty.json")
+    def test_no_findings_clair(self):
+        my_file_handle = open("unittests/scans/clair/clair_empty.json")
         parser = ClairParser()
         findings = parser.get_findings(my_file_handle, None)
         my_file_handle.close()
         self.assertEqual(0, len(findings))
 
-    def test_many_findings(self):
-        my_file_handle = open("unittests/scans/clair/many_vul.json")
+    def test_few_findings_clair(self):
+        my_file_handle = open("unittests/scans/clair/clair_few_vuln.json")
+        parser = ClairParser()
+        findings = parser.get_findings(my_file_handle, None)
+        my_file_handle.close()
+        self.assertEqual(4, len(findings))
+
+    def test_many_findings_clair(self):
+        my_file_handle = open("unittests/scans/clair/clair_many_vul.json")
         parser = ClairParser()
         findings = parser.get_findings(my_file_handle, None)
         my_file_handle.close()
@@ -23,3 +30,24 @@ class TestClairParser(DojoTestCase):
         self.assertEqual("CVE-2018-20839 - (systemd, 237-3ubuntu10.29)", finding.title)
         self.assertEqual(1, len(finding.unsaved_vulnerability_ids))
         self.assertEqual("CVE-2018-20839", finding.unsaved_vulnerability_ids[0])
+
+    def test_parse_no_content_no_findings_clairklar(self):
+        my_file_handle = open("unittests/scans/clair/clairklar_empty.json")
+        parser = ClairParser()
+        findings = parser.get_findings(my_file_handle, None)
+        my_file_handle.close()
+        self.assertEqual(0, len(findings))
+
+    def test_high_findings_clairklar(self):
+        my_file_handle = open("unittests/scans/clair/clairklar_high.json")
+        parser = ClairParser()
+        findings = parser.get_findings(my_file_handle, None)
+        my_file_handle.close()
+        self.assertEqual(6, len(findings))
+
+    def test_mixed_findings_clairklar(self):
+        my_file_handle = open("unittests/scans/clair/clairklar_mixed.json")
+        parser = ClairParser()
+        findings = parser.get_findings(my_file_handle, None)
+        my_file_handle.close()
+        self.assertEqual(6, len(findings))
